Convert to using nbsphinx
