+++ .readthedocs.yml
@@ -9,6 +9,9 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.8"
+  apt_packages:
+    - pandoc
+    # ^ pandoc required by nbsphinx.
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
