+++ setup.cfg
@@ -109,6 +109,7 @@ docs =
     # jupyter included for notebook tutorials.
     jupyter
     myst-parser
+    nbsphinx
     # pyscaffold & pyscaffoldext-markdown included as docs were created with pyscaffoldext-markdown.
     pyscaffold
     pyscaffoldext-markdown >= 0.5
